Rename mapper/synthesize/drift, extract shared types
