+++ clients/admin-ui/src/features/properties/privacy-center-config/form-builder/__tests__/pcShapeToJsonSpec.test.ts
@@ -1,7 +1,7 @@
-import { mapSpecToPcShape } from "../mapper";
-import { synthesizeSpecFromPcShape } from "../synthesize";
+import { jsonSpecToPcShape } from "../jsonSpecToPcShape";
+import { pcShapeToJsonSpec } from "../pcShapeToJsonSpec";
 
-describe("synthesizeSpecFromPcShape", () => {
+describe("pcShapeToJsonSpec", () => {
   it("creates a Form-rooted spec with one element per field", () => {
     const pcShape = {
       email: { label: "Email", field_type: "text" as const, required: true },
@@ -13,7 +13,7 @@ describe("synthesizeSpecFromPcShape", () => {
       },
     };
 
-    const spec = synthesizeSpecFromPcShape(pcShape);
+    const spec = pcShapeToJsonSpec(pcShape);
 
     expect(spec.root).toBe("form");
     expect(spec.elements.form.type).toBe("Form");
@@ -42,8 +42,8 @@ describe("synthesizeSpecFromPcShape", () => {
       },
     };
 
-    const spec = synthesizeSpecFromPcShape(pcShape);
-    const back = mapSpecToPcShape(spec);
+    const spec = pcShapeToJsonSpec(pcShape);
+    const back = jsonSpecToPcShape(spec);
 
     expect(back.errors).toEqual([]);
     expect(back.droppedFeatures).toEqual([]);
@@ -69,7 +69,7 @@ describe("synthesizeSpecFromPcShape", () => {
     };
     const fieldOrder = ["email", "reason", "name", "topics"];
 
-    const spec = synthesizeSpecFromPcShape(pcShape, identityInputs, fieldOrder);
+    const spec = pcShapeToJsonSpec(pcShape, identityInputs, fieldOrder);
 
     expect(spec.elements.form.children).toEqual([
       "f_email",
@@ -77,7 +77,7 @@ describe("synthesizeSpecFromPcShape", () => {
       "f_name",
       "f_topics",
     ]);
-    const back = mapSpecToPcShape(spec);
+    const back = jsonSpecToPcShape(spec);
     expect(back.fieldOrder).toEqual(["email", "reason", "name", "topics"]);
   });
 
@@ -91,7 +91,7 @@ describe("synthesizeSpecFromPcShape", () => {
       name: "optional" as const,
     };
 
-    const spec = synthesizeSpecFromPcShape(pcShape, identityInputs);
+    const spec = pcShapeToJsonSpec(pcShape, identityInputs);
 
     // Canonical legacy order: name → email → phone → customs.
     expect(spec.elements.form.children).toEqual([
@@ -115,7 +115,7 @@ describe("synthesizeSpecFromPcShape", () => {
     const identityInputs = { email: "required" as const };
     const fieldOrder = ["reason", "email"]; // `topics` configured but absent from order
 
-    const spec = synthesizeSpecFromPcShape(pcShape, identityInputs, fieldOrder);
+    const spec = pcShapeToJsonSpec(pcShape, identityInputs, fieldOrder);
 
     expect(spec.elements.form.children).toEqual([
       "f_reason",
@@ -131,7 +131,7 @@ describe("synthesizeSpecFromPcShape", () => {
     const identityInputs = { email: "required" as const };
     const fieldOrder = ["email", "ghost_field", "reason"];
 
-    const spec = synthesizeSpecFromPcShape(pcShape, identityInputs, fieldOrder);
+    const spec = pcShapeToJsonSpec(pcShape, identityInputs, fieldOrder);
 
     expect(spec.elements.form.children).toEqual(["f_email", "f_reason"]);
   });
@@ -153,8 +153,8 @@ describe("synthesizeSpecFromPcShape", () => {
     };
     const fieldOrder = ["topics", "email", "phone", "reason", "name"];
 
-    const spec = synthesizeSpecFromPcShape(pcShape, identityInputs, fieldOrder);
-    const back = mapSpecToPcShape(spec);
+    const spec = pcShapeToJsonSpec(pcShape, identityInputs, fieldOrder);
+    const back = jsonSpecToPcShape(spec);
 
     expect(back.fieldOrder).toEqual(fieldOrder);
   });
